feat(notif): notif on friend connection
